Add tower images
