+++ tower-defense/.namuibundle
@@ -1,0 +1,1 @@
+asset:asset
