+++ tower-defense/src/game_state/tower/mod.rs
@@ -1,9 +1,11 @@
+mod render;
 mod skill;
 
-use crate::card::{Rank, Suit};
-
 use super::*;
+use crate::card::{Rank, Suit};
 use namui::*;
+pub use render::tower_animation_tick;
+use render::{Animation, AnimationKind};
 pub use skill::*;
 use std::{
     fmt::Display,
@@ -18,7 +20,7 @@ pub struct Tower {
     template: TowerTemplate,
     pub status_effects: Vec<TowerStatusEffect>,
     pub skills: Vec<TowerSkill>,
-    animation: Animation,
+    pub(self) animation: Animation,
 }
 impl Tower {
     pub fn new(template: &TowerTemplate, left_top: MapCoord) -> Self {
@@ -92,30 +94,6 @@ impl Tower {
         )
     }
 }
-impl Component for &Tower {
-    fn render(self, ctx: &RenderCtx) {
-        let animation_name = match self.animation.kind {
-            AnimationKind::Idle1 => "idle1",
-            AnimationKind::Idle2 => "idle2",
-            AnimationKind::Attack => "attack",
-        };
-        let image = ctx.image(ResourceLocation::bundle(format!(
-            "tower/{}/{animation_name}.jpg",
-            self.kind.asset_id(),
-        )));
-
-        if let Some(Ok(image)) = image.as_ref() {
-            ctx.add(namui::image(ImageParam {
-                rect: Rect::from_xy_wh(Xy::zero(), image.info.wh()),
-                image: image.clone(),
-                style: ImageStyle {
-                    fit: ImageFit::None,
-                    paint: None,
-                },
-            }));
-        }
-    }
-}
 impl Deref for Tower {
     type Target = TowerTemplate;
 
@@ -139,6 +117,7 @@ pub struct TowerTemplate {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum TowerKind {
+    Barricade,
     High,
     OnePair,
     TwoPair,
@@ -153,10 +132,23 @@ pub enum TowerKind {
 
 impl TowerKind {
     fn asset_id(&self) -> &'static str {
-        todo!()
+        match self {
+            TowerKind::Barricade => "barricade",
+            TowerKind::High => "high",
+            TowerKind::OnePair => "one_pair",
+            TowerKind::TwoPair => "two_pair",
+            TowerKind::ThreeOfAKind => "three_of_a_kind",
+            TowerKind::Straight => "straight",
+            TowerKind::Flush => "flush",
+            TowerKind::FullHouse => "full_house",
+            TowerKind::FourOfAKind => "four_of_a_kind",
+            TowerKind::StraightFlush => "straight_flush",
+            TowerKind::RoyalFlush => "royal_flush",
+        }
     }
     pub fn default_damage(&self) -> usize {
         match self {
+            Self::Barricade => 0,
             Self::High => 5,
             Self::OnePair => 25,
             Self::TwoPair => 50,
@@ -176,6 +168,7 @@ impl Display for TowerKind {
             f,
             "{}",
             match self {
+                Self::Barricade => "Barricade",
                 Self::High => "High",
                 Self::OnePair => "One Pair",
                 Self::TwoPair => "Two Pair",
@@ -222,59 +215,4 @@ pub fn tower_cooldown_tick(game_state: &mut GameState, dt: Duration) {
             tower.cooldown -= cooldown_sub;
         }
     });
-}
-
-pub fn tower_animation_tick(game_state: &mut GameState, now: Instant) {
-    game_state.towers.iter_mut().for_each(|tower| {
-        let animation = &mut tower.animation;
-
-        if now - animation.start_at < animation.duration() {
-            return;
-        }
-
-        animation.transition(match animation.kind {
-            AnimationKind::Idle1 => AnimationKind::Idle2,
-            AnimationKind::Idle2 => AnimationKind::Idle1,
-            AnimationKind::Attack => AnimationKind::Idle1,
-        });
-    });
-}
-
-struct Animation {
-    kind: AnimationKind,
-    start_at: Instant,
-}
-
-impl Animation {
-    fn new() -> Self {
-        Self {
-            kind: AnimationKind::Idle1,
-            start_at: Instant::now(),
-        }
-    }
-
-    fn transition(&mut self, kind: AnimationKind) {
-        self.kind = kind;
-        self.start_at = Instant::now();
-    }
-
-    fn duration(&self) -> Duration {
-        self.kind.duration()
-    }
-}
-
-enum AnimationKind {
-    Idle1,
-    Idle2,
-    Attack,
-}
-
-impl AnimationKind {
-    fn duration(&self) -> Duration {
-        match self {
-            Self::Idle1 => Duration::from_secs(1),
-            Self::Idle2 => Duration::from_secs(1),
-            Self::Attack => Duration::from_secs(1),
-        }
-    }
 }
